Refactor generator modelinfo
