+++ bin/objectbox_model_generator/lib/src/merge.dart
@@ -1,4 +1,4 @@
-import "model.dart";
+import "modelinfo/modelinfo.dart";
 
 Map<String, dynamic> merge(Map<String, dynamic> modelInfo, Map<String, dynamic> annotatedModel) {
   print(annotatedModel);
